Fix points
